chore: remove the reference to the Openfort plugin, and move relevant code into the BeamSDK module External folder
The Openfort plugin only exposed the OpenfortComponent actor component.
To gain direct access the required KeyPair methods, the relevant code was moved to an External folder of the BeamSDK module, along with the LICENSE file that must be provided with the code.
The OpenfortComponent was also converted to OpenfortClient, as a simple UObject, if needed at some point.