feat: more progress on the BeamClient (WIP)
Fixed up the method signatures and return values (currently setup to return TFuture types when the async methods are fleshed out.
Moved the function bodies from the header to the cpp.
More port work, especially related to CreateSessionAsync and SignOperationAsync, as those will be the focus of the demo.
Everything is now roughly in place and method documentation has been updated to keep the header relatively clean.